+++ src/editor/actions/server-save-actions.js
@@ -5,7 +5,8 @@ import {
 } from "../../shared/server-api/notebook";
 import { clearLocalAutosave } from "../tools/local-autosave";
 import { getNotebookID, getRevisionID } from "../tools/server-tools";
-import { updateIomdContent, updateTitle, updateNotebookInfo } from "./actions";
+import { updateTitle, updateNotebookInfo } from "./notebook-actions";
+import { updateIomdContent } from "./editor-actions";
 
 function updateServerSaveStatus(error) {
   let serverSaveStatus;
@@ -54,6 +55,7 @@ export function createNewNotebookOnServer() {
           revision_id: notebook.latest_revision.id,
           revision_is_latest: true,
           serverSaveStatus: "OK",
+          tryItMode: false,
           user_can_save: true,
           username: state.userData.name // in case this notebook was forked
         })