+++ src/editor/state-schemas/language-definitions.js
@@ -3,12 +3,9 @@
 export const jsLanguageDefinition = {
   pluginType: "language",
   languageId: "js",
-  displayName: "Javascript",
-  codeMirrorMode: "javascript",
-  codeMirrorModeLoaded: true,
+  displayName: "JavaScript",
   module: "window",
   evaluator: "eval",
-  keybinding: "j",
   url: ""
 };
 
@@ -19,8 +16,6 @@ const PYODIDE_URL = process.env.USE_LOCAL_PYODIDE
 const pyLanguageDefinition = {
   languageId: "py",
   displayName: "Python",
-  codeMirrorMode: "python",
-  keybinding: "p",
   url: PYODIDE_URL,
   module: "pyodide",
   evaluator: "runPython",