+++ src/editor/style/z-index-styles.js
@@ -3,7 +3,3 @@ export const LAYOUT_MANAGER_IN_BACK_ZINDEX = -10;
 
 export const MODAL_ZINDEX = 1300;
 export const NESTED_MODAL_ZINDEX = 1400;
-
-// FIXME: ideally we will eventually migrate these to css-in-js
-// export const NOTEBOOK_HEADER_ZINDEX = 1000
-// export const EVAL_FRAME_ZINDEX = 0