+++ src/eval-frame/index.jsx
@@ -5,16 +5,12 @@ import { render } from "react-dom";
 // external styles
 import "font-awesome/css/font-awesome.css";
 import "opensans-npm-webfont/style.css";
-import "react-table/react-table.css";
 import "./../../node_modules/katex/dist/katex.min.css";
 
 // iodide styles
-import "./style/eval-container.css";
-import "./style/side-panes.css";
 import "./style/default-presentation.css";
 
 import CSSCascadeProvider from "../shared/components/css-cascade-provider";
-import { initializeDefaultKeybindings } from "./keybindings";
 import initializeUserVariables from "./initialize-user-variables";
 import EvalContainer from "./components/eval-container";
 import ViewModeStylesHandler from "./components/view-mode-styles-handler";
@@ -28,7 +24,6 @@ import "./port-to-editor";
 messagePasserEval.connectDispatch(store.dispatch);
 window.iodide = iodide;
 
-initializeDefaultKeybindings();
 // initialize variables available to the user in an empty notebook, such as
 // the iodide API.
 render(